Added serverdatabase.py that provides a wrapper around the SQlite module for the Server to utilise.
The wrapper consists of four API functions for initialising the database, setting up the sensor tables, inserting and selecting sensor data from the database. The selecting sensor data API allows for selecting data between two times (so a range of data can be retrieved). This will allow for the easy retrieve of sensor data when bulking up data for transmission.

Basic unit tests are provided in test_serverdatabase.py.