Added database sensor saving to the PDU event functions.
The helper function _save_sensor_data_to_database saves the sensor data passed (in bulk) and then commits the database to file to ensure the sensor data is saved. Not sure if this is the most optimal way to do it (as saving this frequently could be performance heavy (?)). Someone to look into if performance becomes an issue.