An initial commit of the RESTful functionality implemented in restful.py within Restful. Basic units tests have been added along side the initial implementation in test_restful.py.
The RESTful implementation is built upon the websocket library https://websockets.readthedocs.io/en/stable/intro.html. The API of restful.py is similar to the websocket library where a serve function is used to register a callable function which is invoked when RESTful requests are received to be handled. The RestfulRequest is a subclass which encapsulates the request for the callable function along with a respond function which the callable function will use to respond to the request.