Initial implementation of the caremulator.py. The emulator reads the configuration in caremulator_config.xml which contains the configurations for each sensor and the emulation formula for that sensor.
Each sensor emulator currently has access to an "x" variable which increments each time the emulation is performed. This allows for the emulation to produce different values each time it is used to be inserted into the emulation database as dummy sensor data. The emulator simply loops through each sensor every delay and inserts the eval formula into the database.