Updated the server to use the XML configuration file config.xml (which incorporates the contents of the previous emulation_config).
The rest of the modules has been updated with the helper function _parse_configuration which parses the config.xml file and extracts the relevant configuration parameters for that module. The command-line arguments have been removed.

The config.xml file is sectioned into tags for the major modules of the server: RESTful, XBee, server, and emulation.