Implemented the new `keep_alive` configuration (under the `RESTful` configuration section of `config.xml`) and into the server as a whole.
A small refactor had to be done to implement this. The RESTful callable function that serves the request now returns the response it wishes to respond back to the client with (as apposed to the previous request object having a dedicated respond function). This was done because the RESTful needs to check the _config["keep_alive"] parameter after the websocket send to know to close the websocket or not (which isn't possible in the `RestfulRequest` class instance).