Integrated the sensor configuration from config.xml into the config parse file.
The sensors configurations are parsed in the config parse function `_parse_configuration` and loaded into `self._config["sensors"]`. The previous declarations of the sensors have been removed and the code that access the declarations have been updated to use the `self._config["sensors"]` dictionary.

The tables for the sensor is only created if the sensor is `enabled`.