Added a new schema section to the server's configuration file.
The schema section contains the information on the low-bandwidth protocol schema structure (header, PDUs, and PDU contents). Each field (the data entry in the header or PDU) is made up of properties that define what that field is (name, C datatype, a constant value). The idea is that the fields are in order, so the first field in the PDU is the first entry and last is the last entry. Each PDU has a special frame_id property which corresponds to the header frame_id field (used to determine which PDU to decode from the packet). An enable property is provided as well in case a PDU is wanted to be disabled.