The Core PDU can be decoded from a byte stream successfully.
The first implementation of a working decoding implementation for decoding the core PDU has been complete. A quite refactor of the schema in config.xml had to be done to achieve this. The start byte and frame id are assumed to the first and second bytes of the PDU stream, then the rest of the bytes correspond to the PDU it self. So, the header section of the config has been removed and integrated into the PDUs themselves for simplicity. Each PDU scheme has a name and ID to uniquely identify it (and to create a LUT of PDUs to compare against the PDU stream frame_id field).

The ProtocolPDU sub class has been implemented which encapsulates the PDUs unique fields. This extracts the struct format for unpacking from the PDU config which it then uses in turn to attempt an unpack from the bytes object passed in decode(). The helper function get_fields_values can then be used after decoding to obtain the decoded fields in a dictionary form.

The main event is the decode_to() in the Protocol class which takes the PDU stream byte object and iteratively goes through it and extracts all PDUs inside the stream. Each decoded PDU fields are then passed to the callable handler which can then handle each field.