Implemented all protocol PDUs.
The AIPDU has been removed because it is no longer useful with the RESTful now serving any client whom isn't the car. So, we know only the car is going to use the PDU protocol.

The sequence number has been removed because even if a missing sequence number was detected, we aren't allowed by rules to send data back to the car so we can't flag to resend a sequence number anyway so no reason to keep it.

The length field has been removed because the length is implicit by the number of fields (and their data lengths) in each PDU.