Implemented dynamic schema creation and integrated into the server.
The refactored protocol files are much, much smaller than previously. This commit is too big (should have been smaller commits but I was tired), the code has been documented so refer to code for the exact changes.

A new test PDU "bean" with a new sensor "bean" was added to the schema configuration to test the system. Just by adding this new PDU to the configuration, the new sensor "bean" was successfully extracted from an encoded payload sent through the protocol client socket and inserted into the staging database. The sensor was available to be accessed over the RESTful API as well.