Implemented RESTful status codes.
The returned response to any RESTful request is of the form `{"status": <code>, "result": {...}}`. Where <code> is the HTTP error code for the request.
Implemented are: `200` for success, `400` if the syntax of the request is wrong e.g. `GET beans`, `500` if something goes wrong with the database operations, `501` if the request contains something that isn't implemented (e.g. `PUT` or a filter), and `404` if the wanted sensor in `GET` is not available (in the server config.xml sensor list).