Implemented valid bitfield for the protocol.
The field `valid_bitfield` is required in all PDUs and is used as a bitfield of valid flags so only fields that have their corresponding bit set to `1` will be extracted from the PDU. If the flag is `0`, the field is ignored.

The byte size of the valid_bitfield can be set per PDU based on the size of the PDU (this commit has them all set to H (short)).