Major refactor.
Updated the xml configuration files to .ini and .json files as xml is not very user friendly.
Moved around the files to two folders now helpers and plugins. Helpers are helper functions and plugins are modules that require loading through the function load.

Moved most of the application logic to main rather than in their own module files.